Update make.bash script.

+++ make.bash
@@ -1,21 +1,71 @@
 #!/bin/bash
+
+verbose=false
+execute=false
+nextIsPath=false
+
+function write {
+    if $verbose
+    then
+        echo $1
+    fi
+}
+
+path="."
+
+if [ $# -gt 0 ]
+then
+    for (( i=1; $i<=$#; i=$i+1 ))
+    do
+        if $nextIsPath
+        then
+            path=${!i}
+            nextIsPath=false
+        else
+            if [ "${!i}" == "-v" ]
+            then
+                verbose=true
+            elif [ "${!i}" == "-e" ]
+            then
+                execute=true
+            elif [ "${!i}" == "-p" ]
+            then
+                nextIsPath=true
+            elif [ "${!i}" == "-c" ]
+            then
+                clear
+            elif [ "${!i}" == "-h" ]
+            then
+                echo "jBookLibrary make bash script"
+                echo "  -c clear screen"
+                echo "  -v show executed commands"
+                echo "  -e run game after compile"
+                echo "  -p {PATH} set path where JAR will be moved"
+                exit
+            else
+                echo "Unknow parameter ${!i}"
+                exit;
+            fi
+        fi
+    done
+fi
+
+if $nextIsPath
+then
+    echo "You must send path with -p argument"
+    exit
+fi
+
 if [ `ls -1 | grep class | wc -l` -gt 0 ]
 then
-    echo "> rm *.class"
+    write "> rm *.class"
     rm *.class
 fi
 
 if [ `ls -1 | grep jar | wc -l` -gt 0 ]
 then
-    echo "> rm *.jar"
+    write "> rm *.jar"
     rm *.jar
-fi
-
-if [ $# -gt 0 ]
-then
-    path=$1
-else
-    path="."
 fi
 
 if [ "${path:${#path}-1:1}" == "/" ]
@@ -23,20 +73,30 @@ then
     path=${path:0:${#path}-1}
 fi
 
-echo "> javac jBookLibrary.java"
+write "> javac jBookLibrary.java"
 javac jBookLibrary.java
 
-echo "> jar cfe jBookLibrary.jar jBookLibrary *.class resources/*"
+write "> jar cfe jBookLibrary.jar jBookLibrary *.class resources/*"
 jar cfe jBookLibrary.jar jBookLibrary *.class resources/*
 
-echo "> chmod u+x jBookLibrary.jar"
+write "> chmod u+x jBookLibrary.jar"
 chmod u+x jBookLibrary.jar
 
 if [ "$path" != "." ]
 then
-    echo "> mv jBookLibrary.jar \"$path/jBookLibrary.jar\""
+    realpath=`realpath -s "$path/"`
+
+    write "> mv jBookLibrary.jar $realpath/jBookLibrary.jar"
     mv jBookLibrary.jar "$path/jBookLibrary.jar"
+else
+    realpath=$path
 fi
 
-echo "> rm *.class"
+write "> rm *.class"
 rm *.class
+
+if $execute
+then
+    write "> java -jar $realpath/jBookLibrary.jar"
+    java -jar "$path/jBookLibrary.jar"
+fi
